chore: test for missing branch
In the past, a pull request was merged that did not cover the new
functionality being introduced (it was missing a branch case). Recently
our team had introduced a PR check for drops in code quality. This
is negatively affecting the team's trust in the check, especially since
it creats a soft-block on the PR for a file that was largely untouched
by the change.

This change aims to create a test case that covers the missing branch