fix: assert response assertions first
Part of investigating the flaky functional tests. There are some
assertions being made after the tests are searching for data in
the payloads. This update reverses the logic, so that we can assert
the request is sane before we try to parse the payload for the data
we're looking for