fix: fix flaky join functional tests
We're getting the following err in func test very often. Trace the code a little bit and found that this is caused by a race condition, when the two upstream builds finish around the same time, they will both try to create the downstream build which is setting the parentBuildId as a String. And eventually one of them won and the `parentBuildId` is the id of the last job instead of an array.

```
Step: Then the "JOIN" job is triggered from "PARALLEL1" and "PARALLEL2" - features/workflow.feature:68
12:11:57      Step Definition: features/step_definitions/workflow.js:124
12:11:57      Message:
12:11:57        AssertionError: expected 56372 to be an array
```